==> R/00_all.qmd <==
---
title: "Master document"
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
  progress: false
---

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis_1.qmd >}}

{{< include 06_analysis_2.qmd >}}

{{< include 07_analysis_3.qmd >}}

{{< include 08_analysis_4.qmd >}}

{{< include 09_analysis_5.qmd >}}
