==> R/01_load.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Loading data

## Loading packages

```{r}
#| label: load_packages_1

library("tidyverse")
```

## Loading data files

```{r}
#| label: load_data_1

blood <- read_csv(here::here(file = "data/_raw/blood.csv"))

tonsil <- read_csv(here::here(file = "data/_raw/tonsil.csv"))

thymus <- read_csv(here::here(file = "data/_raw/thymus.csv"))
```

## Writing data files

```{r}
#| label: write_data_1

write_csv(x = blood,
          file = here::here("data/blood_load.csv"))

write_csv(x = tonsil,
          file = here::here("data/tonsil_load.csv"))

write_csv(x = thymus,
          file = here::here("data/thymus_load.csv"))
```
