==> R/03_augment.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Augmenting data

## Loading packages

```{r}
#| label: load_packages_3

library("tidyverse")

source(file = here::here("R/99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_3

data_clean <- read_csv(file = here::here("data/data_clean.csv"))

PCA_clean <- read_csv(file = here::here("data/PCA_clean.csv"))
```

## Add variables

Here we enrich the dataset with two new variables:

-   hierarchy_level: assigns a numeric level (1–4) describing how specific or mature each cell type is.

-   lineage: groups cells into larger categories such as thymocytes, CD4 T cells, CD8 T cells, T cells, or B cells.

These new variables make it easier to visualize and compare immune-cell maturation patterns across tissues.

```{r}
#| label: add_var_3

level_3 <- c("Tgd", 
             "TCD8", 
             "TCD4", 
             "Bnaive", 
             "BnatEff", 
             "BIgM", 
             "BswMem",
             "Bdn", 
             "B27high", 
             "BnaiveTo", 
             "CC", 
             "CB", 
             "UnswtMem", 
             "SwtMem", 
             "PC")

data_aug <- data_clean |>
  mutate(hierarchy = case_when(cell_type == "Lymphs" ~ 1,
                               cell_type %in% c("T", "B") ~ 2,
                               cell_type %in% level_3 ~ 3,
                               .default = 4),
         .after = cell_type) |>
  add_lineage()
```

The same lineage classification is applied to the PCA dataset so that both data_aug and PCA_aug share identical labels.

```{r}
#| label: add_var_PCA_3

PCA_aug <- PCA_clean |>
  add_lineage()
```

## Writing data files

```{r}
#| label: write_data_3

write_csv(x = data_aug,
          file = here::here("data/data_aug.csv"))

write_csv(x = PCA_aug,
          file = here::here("data/PCA_aug.csv"))
```
