==> R/04_describe.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Describing data

## Loading packages

```{r}
#| label: load_packages_4

library("tidyverse")
library("table1")

source(file = here::here("R/99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_4

data_aug <- read_csv(file = here::here("data/data_aug.csv"))

```

## Description of the clean data set

The data consists of immunological measurements from blood, tonsil and thymus. The dataset consists of `r nrow(data_aug)` observations and `r ncol(data_aug)` variables containing CD markers for the cell types in the specific tissue. An overview of the quantitative markers is given below using a `table1` summary of `data_aug` . **The data is structured in long format,** meaning that each combination of `tissue`, `cell_type` and `CD` appears as one row. This structure makes the dataset suitable for analysis and visualization.

```{r}
#| label: create_table1_4

data_aug |>
  mutate(tissue = factor(tissue,
                         levels = c("thymus", "tonsil", "blood"),
                         labels = c("Thymus", "Tonsil", "Blood"))) |>
  cell_type_order() |>
  table1(~ CVQb + MedQb + PEpos + cell_type | tissue,
         data = _)



```

## Quantifying data set diversity

Using the `summarise()` function, we count the number of unique CD markers and cell types present in the data

```{r}
#| label: count_4

data_aug |>   
  summarise(n_CDs = n_distinct(CD),   
            n_celltypes = n_distinct(cell_type))
```
