==> R/05_analysis_1.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Analysis 1

## Loading packages

```{r}
#| label: load_packages_5

library("tidyverse")
library("broom")
library("ggrepel")
library("patchwork")

source(file = here::here("R/99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_5

PCA_aug <- read_csv(file = here::here("data/PCA_aug.csv"))

```

## Make the PCA fit

The data is split into two data sets, one for MedQb and for PEpos, so we can make a PCA for each.

```{r}

data_MedQb <- PCA_aug |>
  select(tissue, 
         cell_type, 
         lineage, 
         starts_with("MedQb"))

data_PEpos <- PCA_aug |>
  select(tissue, 
         cell_type,
         lineage, 
         starts_with("PEpos"))
```

After all set is split up we use the function `prcomp` was used to make the the PCA.

```{r}
#| label: create_PCA_fit_5

PCA_fit_MedQb <- data_MedQb |> 
  select(where(is.numeric)) |>
  prcomp(scale = TRUE)

PCA_fit_PEpos <- data_PEpos |> 
  select(where(is.numeric)) |>
  prcomp(scale = TRUE)
```

## Make the plot

On the plot we have PC1 on the x-axis and PC2 on the y-axis. Usually most variation can be captured with the first two principal components. The points are stratified using tissue type (blood, tonsil, thymus) as the shapes and cell lineage as the color.

```{r}
scatter_plot_MedQb <- PCA_fit_MedQb |>
  PCA_plot(data = data_MedQb,
           subtitle = "MedQb")

scatter_plot_MedQb
```

Then for PEpos:

```{r}
scatter_plot_PEpos <- PCA_fit_PEpos |>
  PCA_plot(data = data_PEpos,
           subtitle = "PEpos")

scatter_plot_PEpos
```

## Rotation matrices

These were made but ultimately it was a dead end.

**MedQb - all variables:**

```{r}
#| label: create_rotplot2_5

PCA_fit_MedQb |> 
  PCA_rotation(prefix = "MedQb_",
               amount = Inf,
               xlimits = c(-0.25, 0.25),
               ylimits = c(-0.25, 0.25),
               subtitle   = "Rotation matrix for all MedQb variables")

```

**MedQb - top 10 variables:**

```{r}
#| label: create_rotplot3_5

PCA_fit_MedQb |>
  PCA_rotation(prefix = "MedQb_",
               amount = 10,
               xlimits = c(-0.25, 0.25),
               ylimits = c(-0.25, 0.25),
               subtitle   = "MedQb")
```

**PEpos - all variables:**

```{r}
#| label: create_rotplot4_5

PCA_fit_PEpos |>
  PCA_rotation(prefix = "PEpos_",
               amount = Inf,
               xlimits = c(-0.2, 0.2),
               ylimits = c(-0.2, 0.2),
               subtitle = "Rotation matrix for all PEpos variables")
```

**PEpos - top 10 variables:**

```{r}
#| label: create_rotplot5_5

PCA_fit_PEpos |>
  PCA_rotation(prefix = "PEpos_",
               amount = 10,
               xlimits = c(-0.2, 0.2),
               ylimits = c(-0.2, 0.2),
               subtitle = "PEpos")
```

## Plot export

```{r}
#| label: export_plots_5

scatter_plot_MedQb <- scatter_plot_MedQb +
    labs(title = "PCA plot stratified on tissue and lineage")

final_plot <- scatter_plot_MedQb / scatter_plot_PEpos

save_plot(plot = final_plot,
          filename = "05_key_plot.png",
          width = 8,
          height = 9)
```
