==> R/06_analysis_2.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Analysis 2

## Loading packages

```{r}
#| label: load_packages_6

library("tidyverse")

source(file = here::here("R/99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_6

data_aug <- read_csv(file = here::here("data/data_aug.csv"))
```

## Creating auxiliary data frame

The **aim** of this visualization is to have a summarized, comprehensive view of all CD markers used in our data set. From it, the user can extract preliminary insight about constitutive, cohort- and tissue-associated markers, which will be further expanded in following analyses.

A dummy variable `tube_bar` (with the value `"TubeBar"`) is added to enable plotting a horizontal bar above the plot for identifying `cell_type` by `tissue`.

```{r}
#| label: create_tubebar_6

data_aug <- data_aug |>
  mutate(tube_bar = "TubeBar")  
```

## Making the heat dot plot

As the same plot will be used many times, albeit with slight alterations, a function called `heat_dot_plot()` has been created separately and placed in `R/99_proj_func.R`. Our data set, `data_aug` was slightly modified for the analysis; new columns with labels for CD number and tissue were added to allow for correct alphabetical display of CD markers using the custom function `CD_order()`. Additionally, `tube_bar` is plotted over the upper x-axis. Discrete intervals for dot size are displayed in the legend for a comprehensive view of the plot, as well as a color gradient bar for fluorescence.

```{r}
#| label: create_plot1_6
#| fig-width: 8
#| fig-height: 20

data_aug |>
  drop_na(MedQb, PEpos) |>
  cell_type_order() |>
  heat_dot_plot()
```

As the plot figure is too large for the presentation, a subset of `data_aug` will be used for showing data points of interest.

```{r}
#| label: create_plot2_6
#| fig-width: 8
#| fig-height: 10

CDs <- c(1,2,3,4,5,6,7,8,9,10, 15, 16, 17, 18, 19, 45, 74, 81, 96, 99)

final_plot <- data_aug |>
  mutate(CD_num = parse_number(CD)) |>
  filter(CD_num %in% CDs) |>
  drop_na(MedQb, PEpos) |>
  cell_type_order() |>
  heat_dot_plot()

final_plot
```

## Export plots

```{r}
#| label: export_plots_6

save_plot(plot = final_plot,
          filename = "06_key_plot.png",
          width = 11,
          height = 7)
```
