==> R/07_analysis_3.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Analysis 3

## Loading packages

```{r}
#| label: load_packages_7

library("tidyverse")
library("scales")
library("patchwork")
library("broom")
library("rlang")

source(here::here("R", "99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_7

data_aug <- read_csv(here::here("data", "data_aug.csv"))

```

## Correlation of MedQb to CVQb and PEpos

The purpose of these plots is to examine and compare the fluorescence signals across key immune cell types and tissues.

### Flourescence of B cells in blood

Firstly, boxplots of `MedQb`, `CVQb` and `PEpos` based on `CD` are created in ascending order of `MedQb` for `B cells` in the `blood`.

```{r}
#| label: B_cells_blood_7
#| fig-height: 10
#| fig-width: 9

p1 <- data_aug |>
  asc_CDs_boxplot(lineage_filter = "B cells",
                  hierarchy_filter = 3,
                  tissue_filter = "blood",
                  measure = MedQb)

p2 <- data_aug |>
  asc_CDs_boxplot(lineage_filter = "B cells",
                  hierarchy_filter = 3,
                  tissue_filter = "blood",
                  measure = CVQb)

p3 <- data_aug |>
  asc_CDs_boxplot(lineage_filter = "B cells",
                  hierarchy_filter = 3,
                  tissue_filter = "blood",
                  measure = PEpos)

p1 / p2 / p3
```

From the plots, it looks like that there might be a correlation between the three variables. When the `MedQb` increases, the `CVQb` seems to decrease. For `PEpos` the correlation does not seem as strong, but it may increase, when `MedQb` increases.

## Applying a linear model

In the next part, the two correlations are assessed by creating a linear model of the two variables, `CVQb` and `PEpos`, against `MedQb`. Additionally, the correlation coefficients are calculated. The logarithmic values are used.

```{r}
#| label: apply_lm_7

columns <- c("tissue",
             "lineage",
             "term_CVQb",
             "estimate_CVQb",
             "term_PEpos",
             "estimate_PEpos")

data_lm <- data_aug |>
  filter(lineage != "T cells",
         !hierarchy %in% c(1, 2)) |>
  group_by(tissue, lineage) |>
  nest() |>
  mutate(model_object_CVQb = map(.x = data,
                   .f = ~lm(formula = log(MedQb) ~ log(CVQb),
                            data = .x)),
         model_object_PEpos = map(.x = data,
                   .f = ~lm(formula = log(MedQb) ~ log(PEpos),
                            data = .x))) |>
  mutate(model_object_tidy_CVQb = map(.x = model_object_CVQb,
                                 .f = ~tidy(.x, 
                                           conf.int = TRUE,
                                           conf.level = 0.95)),
         model_object_tidy_PEpos = map(.x = model_object_PEpos,
                                 .f = ~tidy(.x, 
                                           conf.int = TRUE,
                                           conf.level = 0.95))) |>
  unnest(model_object_tidy_CVQb,
         model_object_tidy_PEpos) |>
  rename(term_CVQb = "term",
         term_PEpos = "term1",
         estimate_CVQb = "estimate",
         estimate_PEpos = "estimate1") |>
  select(columns) |>
  ungroup() |>
  mutate(
    term_CVQb = case_when(
      term_CVQb == "(Intercept)" ~ "intercept",
      term_CVQb != "(Intercept)" ~ "slope"),
    term_PEpos = case_when(
      term_PEpos == "(Intercept)" ~ "intercept",
      term_PEpos != "(Intercept)" ~ "slope")) |>
  pivot_longer(cols = c(term_CVQb, term_PEpos, estimate_CVQb, estimate_PEpos),
                        names_to = c(".value", "measure"),
                        names_pattern = "(.*)_(.*)") |>
  mutate(measure = str_remove(measure, "term_")) |>
  pivot_wider(id_cols = c("tissue", "lineage"),
              names_from = c("measure", "term"),
              values_from = estimate)

```

The correlation coefficient is now calculated and the two data frames are joined.

```{r}
#| label: paramers_7

data_params <- data_aug |>
  filter(!is.na(MedQb), 
         !is.na(CVQb), 
         !is.na(PEpos),
         lineage != "T cells",
         !hierarchy %in% c(1, 2)) |>
  group_by(tissue, lineage) |>
  summarise(CVQb_cor = cor(log(CVQb), log(MedQb)),
            PEpos_cor = cor(log(PEpos), log(MedQb))) |>
  left_join(y = data_lm,
            by = c("tissue", "lineage"),
            relationship = "one-to-one")
```

In the last part, scatter plots with the linear model are created.

```{r}
#| label: plots_7
#| fig-height: 15

p1 <- data_aug |>
  plot_scatter(params = data_params,
               measure = CVQb,
               lineage_filter = "B cells",
               tissue_filter = "blood",
               color = "blue")


p2 <- data_aug |>
  plot_scatter(params = data_params,
               measure = PEpos,
               lineage_filter = "B cells",
               tissue_filter = "blood",
               color = "hotpink")

p3 <- data_aug |>
  plot_scatter(params = data_params,
               measure = CVQb,
               lineage_filter = "B cells",
               tissue_filter = "tonsil",
               color = "blue")

p4 <- data_aug |>
  plot_scatter(params = data_params,
               measure = PEpos,
               lineage_filter = "B cells",
               tissue_filter = "tonsil",
               color = "hotpink")

p5 <- data_aug |>
  plot_scatter(params = data_params,
               measure = CVQb,
               lineage_filter = "CD4 T cells",
               tissue_filter = "blood",
               color = "blue")

p6 <- data_aug |>
  plot_scatter(params = data_params,
               measure = PEpos,
               lineage_filter = "CD4 T cells",
               tissue_filter = "blood",
               color = "hotpink")

p7 <- data_aug |>
  plot_scatter(params = data_params,
               measure = CVQb,
               lineage_filter = "CD8 T cells",
               tissue_filter = "blood",
               color = "blue")

p8 <- data_aug |>
  plot_scatter(params = data_params,
               measure = PEpos,
               lineage_filter = "CD8 T cells",
               tissue_filter = "blood",
               color = "hotpink")

p9 <- data_aug |>
  plot_scatter(params = data_params,
               measure = CVQb,
               lineage_filter = "Thymocytes",
               tissue_filter = "thymus",
               color = "blue")

p10 <- data_aug |>
  plot_scatter(params = data_params,
               measure = PEpos,
               lineage_filter = "Thymocytes",
               tissue_filter = "thymus",
               color = "hotpink")

final_plot <- (p1 + p2) / (p3 + p4) / (p5 + p6) / (p7 + p8) / (p9 + p10) +
  plot_annotation(
    title = "Correlation of the intensity to variation and positivity",
    caption = "cor: correlation coefficient") &
  theme(plot.caption = element_text(size = 11))

final_plot
```

## Plot export

```{r}
#| label: export_plots_7

save_plot(plot = final_plot,
          filename = "07_key_plot.png",
          width = 7.5,
          height = 9)
```
