==> R/08_analysis_4.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Analysis 4

## Loading packages

```{r}
#| label: load_packages_8

library("tidyverse")
library("broom")
library("patchwork")

source(here::here(file = "R/99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_8

data_aug <- read_csv(here::here(file = "data/data_aug.csv"))
```

The aim of this analysis is to explore how different CD markers change across immune cell types and maturation stages in several tissues, using linear models and simple visualizations to highlight meaningful differences.

## Maturation of immune cells

This part prepares the data and runs the linear models. We keep only the detailed cell-type levels we care about and remove some unwanted columns and populations. After cleaning, we group the data by tissue, CD marker, and lineage, and then fit a linear model (`log(MedQb) ~ cell_type`) for each group.\
Once the models are fit, we classify whether each term is significant, rebuild the cell-type names, and adjust estimates by adding the intercept back in so everything stays on the same scale.

```{r}
#| label: apply_lm_8

columns <- c("CD", 
             "tissue",
             "lineage",
             "cell_type",
             "estimate",
             "conf.low", 
             "conf.high",
             "p.value",
             "is_significant")

data_analysis4 <- data_aug |>
  cell_type_order() |>
  filter(hierarchy %in% c(3, 4),
         !cell_type %in% c("Tgd", "TCD4", "TCD8")) |>
  select(-c("CVQb", "PEpos")) |>
  drop_na(MedQb) |>
  group_by(tissue, CD, lineage) |>
  filter(n() > n_distinct(cell_type)) |>
  fit_linear_model(log(MedQb) ~ cell_type)|>
  mutate(is_significant = case_when(term == "(Intercept)" ~ "reference",
                                    p.value < 0.05 ~ "yes",
                                    p.value >= 0.05 ~ "no")) |>
  mutate(
    intercept = estimate[term == "(Intercept)"],
    estimate = case_when(
      term == "(Intercept)" ~ estimate,
      term != "(Intercept)" ~ estimate + intercept),
    conf.low = case_when(
      term == "(Intercept)" ~ conf.low,
      term != "(Intercept)" ~ conf.low + intercept),
    conf.high = case_when(
      term == "(Intercept)" ~ conf.high,
      term != "(Intercept)" ~ conf.high + intercept),
    cell_type = case_when(
      term == "(Intercept)" & tissue == "blood" & lineage == "B cells" ~ "Bnaive",
      term == "(Intercept)" & tissue == "tonsil" ~ "BnaiveTo",
      term == "(Intercept)" & lineage == "CD4 T cells" ~ "TCD4naive", 
      term == "(Intercept)" & lineage == "CD8 T cells" ~ "TCD8naive",
      term == "(Intercept)" & tissue == "thymus"  ~ "DN34p",
    .default = str_remove(term, "cell_type"))) |>
  select(all_of(columns)) |>
  ungroup()
```

The results are then combined into a single table called data_analysis4, which serves as the base for visualization and further interpretation. The function `plot_cd_expression()` is used to visualize these model results, plotting the estimated expression levels and their confidence intervals for each CD marker across tissues and cell types.

## Significant CDs

Here we check which CD markers actually show significant differences between cell types.\
We filter for terms with p \< 0.05, count how many CDs are significant within each lineage.

```{r}
#| label: count_sig_CDs_8

cds_significant <- data_analysis4 |>
  filter(is_significant == "yes") |>      
  group_by(lineage) |>                    
  summarise(n_significant_CDs = n_distinct(CD)) |>
  arrange(desc(n_significant_CDs))        

cds_significant
```

A column plot of the count of significant CDs for each lineage is created below.

```{r}
#| label: plot_sig_CDs_8

sig_CD_count <- data_analysis4 |>
  plot_sig_CDs(by_variable = lineage, 
                 fill_color = "#fa2840")


sig_CD_count
```

## B cells

```{r}
#| label: B_cells_8
#| fig-height: 8

cds_interest_b <- c("CD11a", "CD69", "CD79b", "CD80")

p1 <- data_analysis4 |>
  plot_cd_expression(lineage_filter = "B cells",
                     cds = cds_interest_b,
                     subtitle = "B cells")

p1
```

**B cell markers:** CD11a (cell adhesion), CD69 (early activation), CD79b (B cell receptor signaling), CD80 (co-stimulation).

## T cells

```{r}
#| label: CD4_T_cells_8
#| fig-height: 8

cds_interest_cd4 <- c("CD10", "CD25", "CD4_MEM-241", "CD62L")

p2 <- data_analysis4 |>
  plot_cd_expression(lineage_filter = c("Thymocytes", "CD4 T cells"),
                     cds = cds_interest_cd4,
                     cell_type_exclude = "CD8",
                     subtitle = "CD4 T cells")

p2
```

**CD4 T cell markers:** CD10 (early maturation), CD25 (activation), CD4_MEM-241 (memory), CD62L (migration).

```{r}
#| label: CD8_T_cells_8
#| fig-height: 8

cds_interest_cd8 <- c("CD27", "CD38", "CD71", "CD95")

p3 <- data_analysis4 |>
  plot_cd_expression(lineage_filter = c("Thymocytes", "CD8 T cells"),
                     cds = cds_interest_cd8,
                     cell_type_exclude = "CD4",
                     subtitle = "CD8 T cells")

p3
```

**CD8 T cell markers:** CD27 (memory), CD38 (activation), CD71 (proliferation), CD95 (apoptosis).

## Export plots

```{r}
#| label: export_plots_8

save_plot(plot = sig_CD_count, 
          filename = "08_key_plot_1.png")

p1 <- p1 + 
  labs(title = "Selected significant CDs during maturation stratified on lineage")

final_plot <- p1 + p2 + p3

save_plot(plot = final_plot, 
          filename = "08_key_plot_2.png",
          width = 22,
          height = 9)
```
