==> R/09_analysis_5.qmd <==
---
format:
  html:
    embed-resources: true
project:
  output-dir: results
editor: visual
editor_options: 
  chunk_output_type: console
execute: 
  warning: false
  message: false
---

# Analysis 5

## Loading packages

```{r}
#| label: load_packages_9

library("tidyverse")
library("broom")
library("patchwork")

source(here::here("R", "99_proj_func.R"))
```

## Loading data files

```{r}
#| label: load_data_9

data_aug <- read_csv(file = here::here("data/data_aug.csv"))
```

## Differences in CD makers on CD4 and CD8 T cells

In the data set, the `CD4 T cells` and `CD8 T cells` exist in two similar and parallel subsets in the `thymus`, `SP1ap` and `SP1am`, and four subsets in the `blood`, `naive`, `central memory` (CM) `effector memory` (EM) and `TEMRA`. It is therefore interesting to investigate differences in CD marker expression between `CD4` and `CD8` T cells for each pair of subset.

## Applying a linear model

To assess significant differences in CD marker expression between each pair, a linear model was applied using the `map()` function as described in analysis 4. In order to compare the CDs for a specific subset between `CD4` and `CD8`, the data must be grouped by the pair, hence why the variable `pair` was added. As the `CD4 T cells` are listed first in the function `cell_type_order`, their estimate corresponds the intercept, while the estimate of the `CD8 T cells` is the slope, corresponding to how much the CD expression differs from the `CD4 T cells`. Lastly, the full estimates were calculated by adding the intercept to the estimate of the `CD8 T cells`, and the variable `is_significant` and `any_significant` is added from the calculated p-value. The step `filter(n() > n_distinct(cell_type))` is added to ensure more than one observation for each `cell_type` in each `pair`.

```{r}
#| label: apply_lm_9

columns <- c("tissue", 
             "CD",
             "cell_type",
             "pair",
             "estimate", 
             "conf.low", 
             "conf.high",
             "p.value",
             "is_significant")

data_analysis5  <- data_aug |>
  filter(str_detect(cell_type, "CD4") | str_detect(cell_type, "CD8"),
         cell_type != "CD4ISP",
         cell_type != "TCD8RAdim",
         hierarchy == 4) |>
  select(-c("CVQb", "PEpos")) |>
  drop_na(MedQb) |>
  mutate(pair = case_when(str_detect(cell_type, "SP1ap") ~ "SP1ap",
                          str_detect(cell_type, "SP1am") ~ "SP1am",
                          str_ends(cell_type, "naive") ~ "naive",
                          str_ends(cell_type, "CM") ~ "CM",
                          str_ends(cell_type, "EM") ~ "EM",
                          str_detect(cell_type, "TEMRA") ~ "TEMRA"),
         .before = cell_type) |>
  group_by(tissue, CD, pair) |>
  filter(n() > n_distinct(cell_type)) |>
  fit_linear_model(log(MedQb) ~ cell_type)|>
  mutate(is_significant = case_when(
    p.value[term != "(Intercept)"] < 0.05 ~ "yes",
    p.value[term != "(Intercept)"] >= 0.05 ~ "no")) |>
  mutate(
    intercept = estimate[term == "(Intercept)"],
    estimate = case_when(
      term == "(Intercept)" ~ estimate,
      term != "(Intercept)" ~ estimate + intercept),
    conf.low = case_when(
      term == "(Intercept)" ~ conf.low,
      term != "(Intercept)" ~ conf.low + intercept),
    conf.high = case_when(
      term == "(Intercept)" ~ conf.high,
      term != "(Intercept)" ~ conf.high + intercept),
    cell_type = case_when(
      term == "(Intercept)" & pair == "SP1ap" ~ "CD4SP1ap",
      term == "(Intercept)" & pair == "SP1am" ~ "CD4SP1am",
      term == "(Intercept)" & pair == "naive" ~ "TCD4naive",
      term == "(Intercept)" & pair == "CM" ~ "TCD4CM",
      term == "(Intercept)" & pair == "EM" ~ "TCD4EM", 
      term == "(Intercept)" & pair == "TEMRA" ~ "TCD4TEMRA",
      .default = str_remove(term, "cell_type"))) |>
  select(all_of(columns)) |>
  ungroup() |>
  mutate(lineage = case_when(str_detect(cell_type, "CD4") ~ "CD4 T cells",
                             .default = "CD8 T cells"),
         .before = cell_type)

```

## Count of significant CDs

Below are the number of significant CDs for each pair calculated and visualized.

```{r}
#| label: count_sig_CDs_9

sig_CD_count <- data_analysis5 |>
  plot_sig_CDs(by_variable = pair, 
                 fill_color = "#fa2840")


sig_CD_count
```

## Visualization of significant CDs between each pair

The significant CDs for each pair are visualized with an error bar plot of the estimate and the confidence intervals by using the function `plot_CD4vsCD8()`. The six figures are plotted together by using the package `patchwork`.

```{r}
#| label: create_plots_9
#| fig-height: 10
#| fig-width: 8

p1 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair_filter = "SP1ap",
                legend_position = "none")

p2 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair_filter = "SP1am",
                legend_position = "none")

p3 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair_filter = "naive",
                legend_position = "none")

p4 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair = "CM",
                legend_position = "none")

p5 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair_filter = "EM",
                legend_position = "none")

p6 <- data_analysis5 |>
  CD_order() |>
  plot_CD4vsCD8(pair_filter = "TEMRA",
                legend_position = "bottom")

(p1 + p2) / (p3 + p4) / (p5 + p6) +
  plot_annotation(
    title = "Significant CD Markers Differentiating CD4 and CD8 T Cell Subsets") &
  theme(plot.title = element_text(size = 18))
```

## Overview of CDs

```{r}
#| label: CDs_overview_9

final_plot <- data_analysis5 |>
  CD_order() |>
  group_by(CD)|>
  filter(any(is_significant == "yes")) |>
  ungroup() |>
  group_by(CD, pair) |>
  mutate(highest = if_else(
    estimate[str_detect(cell_type, "CD8")] >
    estimate[str_detect(cell_type, "CD4")], 
    "CD8 T cells",
    "CD4 T cells")) |>
  mutate(pair = factor(x = pair, 
                       levels = c("TEMRA",
                                  "EM",
                                  "CM",
                                  "naive",
                                  "SP1am",
                                  "SP1ap"))) |>
  ggplot(aes(x = CD,
             y = pair)) +
  geom_tile(aes(fill = is_significant),
            alpha = 0.4,
            color = "white") +
  geom_point(aes(shape = highest,
                 color = highest,),
             size = 2) +
  scale_fill_manual(values = c("no" = "#fa2840",
                                "yes" = "#29cf8f")) +
  scale_color_manual(values = c("CD4 T cells" = "black",
                               "CD8 T cells" = "#FFD700")) +
  scale_shape_manual(values = c("CD4 T cells" = 16, 
                                "CD8 T cells" = 15)) +
  theme_bw(base_size = 14) +
  theme(axis.text.x = element_text(angle = 90, 
                                    vjust = 0.5, 
                                    hjust = 1),
        legend.position = "bottom",
        legend.box = "vertical") +
  labs(title = "Significant CDs for each pair",
       fill = "Is significant",
       shape = "Highest log(MedQb)",
       color = "Highest log(MedQb)",
       x = "CD",
       y = "Pair")

final_plot
```

## Plot export

```{r}
#| label: export_plots_9

save_plot(plot = sig_CD_count,
          filename = "09_key_plot_1.png")
  
save_plot(plot = final_plot,
          filename = "09_key_plot_2.png",
          width = 10,
          height = 5)
```
